New system let's go!

+++ BizHawk.Emulation.Cores/Consoles/Sega/SMS/SMS.IDebuggable.cs
@@ -144,7 +144,7 @@ namespace BizHawk.Emulation.Cores.Sega.MasterSystem
 
 		public int TotalExecutedCycles
 		{
-			get { return Cpu.TotalExecutedCycles; }
+			get { return (int)Cpu.TotalExecutedCycles; }
 		}
 	}
 }
